.claude/worktrees/xenodochial-dubinsky/requirements.txt
# ============================================================================
# AI JOB FOUNDRY - DEPENDENCIES
# Python 3.11+
# ============================================================================

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Web Scraping & Automation
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Google APIs
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AI & LLM
openai==1.3.7
anthropic==0.7.7
litellm==1.0.3

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
colorama==0.4.6
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Email
email-validator==2.1.0

# Logging
loguru==0.7.2

# Database (optional)
sqlite3-python==1.0.0

# File handling
pypdf2==3.0.1
python-docx==1.1.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
